+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 * **compact:** remove compacting spinner once compact completes ([#66](https://github.com/sendbird-playground/stave/issues/66)) ([5f413d2](https://github.com/sendbird-playground/stave/commit/5f413d21de6f6277787e21a86ad5835f1d871a9f))
 * improve inline auto-complete behavior ([#65](https://github.com/sendbird-playground/stave/issues/65)) ([c3e4e74](https://github.com/sendbird-playground/stave/commit/c3e4e743193330c0551b8db2e500ee3bce2303cd))
+* **skills:** handle skill-only invocations ([#68](https://github.com/sendbird-playground/stave/issues/68)) ([fe2694f](https://github.com/sendbird-playground/stave/commit/fe2694fc3d81c8d553edd3c103b348f2f74bda92))
 ## [0.0.26](https://github.com/sendbird-playground/stave/compare/v0.0.25...v0.0.26) (2026-03-29)
 
 ### Features